Rewrite relay satellite launch controller in terms of EventDriver.

+++ smallship/launchcontroller-header.cs
@@ -1,2 +1,2 @@
 // LaunchController
-const int LAUNCH_BURN_DURATION = 10; // In seconds
+const double LAUNCH_BURN_DURATION = 10.0; // In seconds
